Change sqlite3 db to memory mode.

+++ util/data_cache.py
@@ -1,4 +1,5 @@
-import os, fnmatch, re, threading, sublime, json, sqlite3, shutil, time
+import os, fnmatch, re, threading, sublime, sqlite3, shutil, time
+from multiprocessing.pool import ThreadPool
 from .settings import get_settings_param, GLOBAL_SET
 
 CREATE_LIBS_INFO_SQL = '''
@@ -9,12 +10,14 @@ create table if not exists libs_info (
     primary key(id)
 );
 '''
+
 INSERT_FOLDER_INFO = '''
 replace into libs_info(id, parent_id, folder) values
 (?, ?, ?);
 '''
+
 QUERY_FOLDER = '''
-select id from libs_info where folder = ?;
+select id, parent_id from libs_info where folder = ?;
 '''
 
 CREATE_LIBS_SQL = '''
@@ -25,13 +28,17 @@ create table if not exists libs (
     param_len tinyint(2) not null,
     row_num int unsigned not null,
     completion varchar(256) not null,
-    primary key(id, fun_name, param_len)
+    primary key(id, mod_name, fun_name, param_len)
 ); 
 '''
 
 INSERT_LIBS_SQL = '''
 replace into libs(id, mod_name, fun_name, param_len, row_num, completion) values 
 (?, ?, ?, ?, ?, ?);
+'''
+
+DEL_LIBS_SQL = '''
+delete from libs where id = ? and mod_name = ?;
 '''
 
 QUERY_COMPLETION = '''
@@ -47,39 +54,33 @@ select folder, fun_name, param_len, row_num from libs join libs_info where libs_
 '''
 
 class DataCache:
-    def __init__(self, dir = '', data_type = '', cache_dir = ''):
+    def __init__(self, data_type = '', cache_dir = '', dir = None):
         self.dir = dir
         self.data_type = data_type
         self.cache_dir = cache_dir
         self.re_dict = GLOBAL_SET['compiled_re']
-        self.version = get_settings_param('sublime_erlang_version', '0.0.0')
-        self.is_ready = False
+        self.pool_size = 8
+        self.folder_id = 1
         if cache_dir != '':
             self.__init_db()
 
     def __init_db(self):
-        db_path = self.__get_filepath('completion')
-        try:
-            self.db_con = sqlite3.connect(db_path, check_same_thread = False)
-            self.db_cur = self.db_con.cursor()
-            self.db_cur.execute(CREATE_LIBS_INFO_SQL)
-            self.db_cur.execute(CREATE_LIBS_SQL)
-        except Exception as e:
-            self.db_con.close()
-            print('Exception', e)
+        if os.path.exists(self.cache_dir): 
             shutil.rmtree(self.cache_dir)
-            print('Remove dir {}.'.format(self.cache_dir))
-            db_path = self.__get_filepath('completion')
-            self.db_con = sqlite3.connect(db_path, check_same_thread = False)
-            self.db_cur = self.db_con.cursor()
-            self.db_cur.execute(CREATE_LIBS_SQL)
+        # self.db_con = sqlite3.connect(':memory:', check_same_thread = False)
+        self.db_con = sqlite3.connect('D:/{}'.format(self.data_type), check_same_thread = False)
+        self.db_cur = self.db_con.cursor()
+        self.db_cur.execute(CREATE_LIBS_INFO_SQL)
+        self.db_cur.execute(CREATE_LIBS_SQL)
 
     def query_mod_fun(self, module):
-        if not self.is_ready:
-            return []
-
-        self.db_cur.execute(QUERY_COMPLETION, (module, ))
-        query_data = self.db_cur.fetchall()
+        query_data = []
+        try:
+            self.lock.acquire(True)
+            self.db_cur.execute(QUERY_COMPLETION, (module, ))
+            query_data = self.db_cur.fetchall()
+        finally:
+            self.lock.release()
 
         completion_data = []
         for (fun_name, param_len, param_str) in query_data:
@@ -90,11 +91,13 @@ class DataCache:
         return completion_data
 
     def query_all_mod(self):
-        if not self.is_ready:
-            return []
-        
-        self.db_cur.execute(QUERY_ALL_MOD)
-        query_data = self.db_cur.fetchall()
+        query_data = []
+        try:
+            self.lock.acquire(True)
+            self.db_cur.execute(QUERY_ALL_MOD)
+            query_data = self.db_cur.fetchall()
+        finally:
+            self.lock.release()
 
         completion_data = []
         for (mod_name, ) in query_data:
@@ -103,11 +106,13 @@ class DataCache:
         return completion_data
 
     def query_fun_position(self, module, function):
-        if not self.is_ready:
-            return []
-
-        self.db_cur.execute(QUERY_POSITION, (module, function))
-        query_data = self.db_cur.fetchall()
+        query_data = []
+        try:
+            self.lock.acquire(True)
+            self.db_cur.execute(QUERY_POSITION, (module, function))
+            query_data = self.db_cur.fetchall()
+        finally:
+            self.lock.release()
 
         completion_data = []
         for (folder, fun_name, param_len, row_num) in query_data:
@@ -116,7 +121,7 @@ class DataCache:
 
         return completion_data
 
-    def build_module_dict(self, filepath, folder_id):
+    def build_module_index(self, filepath, folder_id):
         with open(filepath, encoding = 'UTF-8', errors='ignore') as fd:
             content = fd.read()
             code = re.sub(self.re_dict['comment'], '\n', content)
@@ -138,13 +143,16 @@ class DataCache:
                     param_len = len(param_list)
                     if (fun_name, param_len) in export_fun:
                         del(export_fun[(fun_name, param_len)])
-                        self.db_cur.execute(INSERT_LIBS_SQL, (folder_id, module, fun_name, param_len, row_num, param_str))
+                        try:
+                            self.lock.acquire(True)
+                            self.db_cur.execute(INSERT_LIBS_SQL, (folder_id, module, fun_name, param_len, row_num, param_str))
+                        finally:
+                            self.lock.release()
                 row_num += 1
 
     def get_module_from_path(self, filepath):
         (path, filename) = os.path.split(filepath)
         (module, extension) = os.path.splitext(filename)
-
         return module
 
     def format_param(self, param_str):
@@ -162,48 +170,66 @@ class DataCache:
         completion = '{0}({1})${2}'.format(funname, param_str, len + 1)
         return completion
 
-    def __get_filepath(self, filename):
-        if not os.path.exists(self.cache_dir): 
-            os.makedirs(self.cache_dir)
-        real_filename = '{0}_{1}'.format(self.data_type, filename)
-        filepath = os.path.join(self.cache_dir, real_filename)
-        return filepath
-
-    def __dump_json(self, filename, data):
-        filepath = self.__get_filepath(filename)
-        with open(filepath, 'w') as fd:
-            fd.write(json.dumps(data))
-
     def build_data(self):
         all_filepath = []
-        folder_id = 1
-        for dir in self.dir:
-            if self.is_build_index(dir):
-                return
+        start_time = time.time()
+        task_pool = ThreadPool(self.pool_size)
+        self.lock = threading.Lock()
 
-            self.db_cur.execute(INSERT_FOLDER_INFO, (folder_id, 0, dir))
-            parent_id = folder_id
-            folder_id += 1
-            for root, dirs, files in os.walk(dir):
-                self.db_cur.execute(INSERT_FOLDER_INFO, (folder_id, parent_id, root))
-                for file in fnmatch.filter(files, '*.erl'):
-                    self.build_module_dict(os.path.join(root, file), folder_id)
-                folder_id += 1
+        if self.dir == None:
+            folders = self.get_all_open_folders()
+        else:
+            folders = self.dir
+
+        is_save_build_index = False
+        for folder in folders:
+            if self.get_folder_id(folder) != None:
+                continue
+
+            print('build {}: {} index'.format(self.data_type, folder))
+            is_save_build_index = True
+            self.db_cur.execute(INSERT_FOLDER_INFO, (self.folder_id, 0, folder))
+            parent_id = self.folder_id
+            for root, dirs, files in os.walk(folder):
+                erl_files = fnmatch.filter(files, '*.erl')
+                if erl_files == []:
+                    continue
+                    
+                self.db_cur.execute(INSERT_FOLDER_INFO, (self.folder_id, parent_id, root))
+                for file in erl_files:
+                    all_filepath.append((os.path.join(root, file), self.folder_id))
+                self.folder_id += 1
         
+        task_pool.starmap(self.build_module_index, all_filepath)
         self.db_con.commit()
+        is_save_build_index and print("build {} index, use {} second".format(self.data_type, time.time() - start_time))
 
-    def is_build_index(self, folder):
+    def get_all_open_folders(self):
+        all_folders = []
+        for window in sublime.windows():
+            all_folders = all_folders + window.folders()
+
+        return all_folders
+
+    def get_folder_id(self, folder):
         self.db_cur.execute(QUERY_FOLDER, (folder, ))
-        return self.db_cur.fetchall() != []
+        for (fid, pid) in self.db_cur.fetchall():
+            return (fid, pid)
+        return None
+
+    def rebuild_module_index(self, filepath):
+        (folder, filename) = os.path.split(filepath)
+        (module, extension) = os.path.splitext(filename)
+        (fid, pid) = self.get_folder_id(folder)
+        print('rebuild_module_index', folder, fid, pid)
+        self.db_cur.execute(DEL_LIBS_SQL, (fid, module))
+        self.build_module_index(filepath, fid)
+        self.db_con.commit()
 
     def build_data_async(self):
         this = self
         class BuildDataAsync(threading.Thread):
             def run(self):
-                start_time = time.time()
-                print("start", start_time)
                 this.build_data()
-                print("end", time.time() - start_time)
-                this.is_ready = True
                 
         BuildDataAsync().start()
